requirements.txt
beautifulsoup4==4.11.1
boto3==1.28.44
botocore==1.31.44
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
importlib-metadata==5.2.0
jmespath==1.0.1
lxml==5.3.1
mysqlclient==2.1.1
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
requests-toolbelt==0.10.1
s3transfer==0.6.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
typing-extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
